+++ attributes/default.rb
@@ -1,0 +1,2 @@
+default['domain'] = "priv.evertrue.com"
+default['fqdn'] = "#{node.name}.#{node['domain']}"