+++ recipes/default.rb
@@ -7,18 +7,30 @@
 # All rights reserved - Do Not Redistribute
 #
 
-template "/etc/hosts" do
+require 'socket'
+
+t = template "/etc/hosts" do
   source "hosts.erb"
   owner "root"
   group "root"
   mode 0644
   backup false
+  action :nothing
 end
+t.run_action(:create)
 
-execute "Set current hostname" do
-  command "hostname #{node.name}"
+e = execute "Set current hostname" do
+  command "hostname #{node['fqdn']}"
+  only_if { Socket.gethostname != node['fqdn'] }
+  action :nothing
+end
+e.run_action(:run)
+
+ohai "reload_hostname" do
+  plugin "hostname"
+  only_if { Socket.gethostname != node['fqdn'] }
 end
 
 execute "Set permanent hostname" do
-  command "sysctl kernel.hostname=#{node.name}"
+  command "sysctl kernel.hostname=#{node['fqdn']}"
 end