+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file is used to list changes made in each version of hostname.
 
+## 1.0.1:
+
+* Total refactor
+* hard-connect node FQDN to node.name
+* hard-code node domain to "priv.evertrue.com"
+
 ## 0.1.0:
 
 * Initial release of hostname